Rename MeeseeksBoxFactory to BackgroundTaskManagerFactory
