+++ runtime/src/commonMain/kotlin/dev/mattramotar/meeseeks/runtime/dsl/box/BackgroundTaskManager.kt
@@ -5,16 +5,16 @@ import dev.mattramotar.meeseeks.runtime.AppContext
 import dev.mattramotar.meeseeks.runtime.TaskWorkerRegistry
 import dev.mattramotar.meeseeks.runtime.BackgroundTaskManager
 import dev.mattramotar.meeseeks.runtime.BackgroundTaskConfig
-import dev.mattramotar.meeseeks.runtime.impl.MeeseeksBoxFactory
+import dev.mattramotar.meeseeks.runtime.impl.BackgroundTaskManagerFactory
 
 
-fun MeeseeksBox(
+fun BackgroundTaskManager(
     context: AppContext,
     config: BackgroundTaskConfig = BackgroundTaskConfig(),
     registry: TaskWorkerRegistry
 ): BackgroundTaskManager {
 
-    val factory = MeeseeksBoxFactory()
+    val factory = BackgroundTaskManagerFactory()
 
     return factory.create(
         context,
