+++ runtime/src/commonMain/kotlin/dev/mattramotar/meeseeks/runtime/dsl/registry/TaskWorkerRegistryBuilderExtensions.kt
@@ -4,7 +4,7 @@ import dev.mattramotar.meeseeks.runtime.TaskWorkerRegistry
 import dev.mattramotar.meeseeks.runtime.TaskWorker
 import dev.mattramotar.meeseeks.runtime.Task
 
-object MeeseeksRegistryBuilderExtensions {
+object TaskWorkerRegistryBuilderExtensions {
     fun TaskWorkerRegistry.Builder.register(
         type: String,
         createTaskWorker: (Task) -> TaskWorker
