+++ runtime/src/nativeMain/kotlin/dev/mattramotar/meeseeks/runtime/impl/BackgroundTaskManagerFactory.kt
@@ -4,22 +4,24 @@ import dev.mattramotar.meeseeks.runtime.BackgroundTaskManager
 import dev.mattramotar.meeseeks.runtime.BackgroundTaskConfig
 import dev.mattramotar.meeseeks.runtime.AppContext
 import dev.mattramotar.meeseeks.runtime.TaskWorkerRegistry
+import platform.BackgroundTasks.BGTaskScheduler
 
-internal actual class MeeseeksBoxFactory actual constructor() {
+internal actual class BackgroundTaskManagerFactory {
     actual fun create(
         context: AppContext,
         registry: TaskWorkerRegistry,
         config: BackgroundTaskConfig
     ): BackgroundTaskManager {
         val database = MeeseeksAppDatabase.require(context)
+        val bgTaskScheduler = BGTaskScheduler.sharedScheduler
         val workRequestFactory = WorkRequestFactory()
-        val taskScheduler = TaskScheduler()
+        val taskScheduler = TaskScheduler(database, bgTaskScheduler)
         val taskRescheduler = TaskRescheduler(database, taskScheduler, workRequestFactory)
         return RealBackgroundTaskManager(
-            database = database,
-            workRequestFactory = workRequestFactory,
-            taskScheduler = taskScheduler,
-            taskRescheduler = taskRescheduler
+            database,
+            workRequestFactory,
+            taskScheduler,
+            taskRescheduler
         )
     }
 }
